Add funcs to calculate cov and corr of the signal in a mixed sample.
Implemented sig_cov and sig_corr to estimate the covariance and the
correlation matrices of a signal sample in a mixed sample. The method
works if the fraction of signal in the mixed sample and the covariance
and means of the background. Some more explanation can be found [in
this notebook](https://github.com/tmadlener/phys_utils/blob/master/python/docs/cov_signal_from_mixed.ipynb).