Implement a more sophisticated handling of the .batch step for checking the status.
Previously the status CANCELLED was not handled at all. To handle it properly
it is necessary to check the .batch step, since there the proper exit code can
be found and also for the normal step the info by whom the job has been cancelled
is included, which makes it hard to handle in the way the other statuses are
handled.